refactor(lb): added tests for load balancing module
